Odoo advisory 2016-07b-ogone-eval
Ported patch from 9.0 to 8.6.

The advisory didn't mention 8.0 but I've ported the patch anyway.

If there's a permissions issue when we come to use this, then
this patch is a good candidate to start looking for the bug!

I've done the equivalent under the old field name,
which was to set groups="base.group_system",
just to be sure.